update cli gitignore
