+++ cli/src/main/java/com/github/martingaston/Display.java
