+++ cli/src/main/java/com/github/martingaston/player/PlayerHuman.java
