+++ cli/src/main/java/com/github/martingaston/state/StateCLI.java
