+++ cli/src/main/java/com/github/martingaston/state/StateFile.java
