+++ cli/src/main/java/com/github/martingaston/state/StateUserSelect.java
