+++ cli/src/main/java/com/github/martingaston/storage/Storage.java
