+++ cli/src/test/java/com/github/martingaston/DisplayTest.java
