+++ cli/src/test/java/com/github/martingaston/GameTest.java
