+++ cli/src/test/java/com/github/martingaston/player/CLIPlayersTest.java
@@ -1,6 +1,8 @@
-package com.github.martingaston.tictactoe.player;
+package com.github.martingaston.player;
 
 import com.github.martingaston.tictactoe.board.Board;
+import com.github.martingaston.tictactoe.player.Player;
+import com.github.martingaston.tictactoe.player.Players;
 import org.junit.Before;
 import org.junit.Test;
