+++ cli/src/test/java/com/github/martingaston/player/PlayerHumanTest.java
