+++ tictactoe/src/main/java/com/github/martingaston/tictactoe/state/State.java
@@ -1,4 +1,4 @@
-package com.github.martingaston.tictactoe.state;
+package com.github.martingaston.state;
 
 import com.github.martingaston.tictactoe.GameModes;
 import com.github.martingaston.tictactoe.board.Board;
